+++ account-kit/signer/src/index.ts
@@ -20,3 +20,4 @@ export { AlchemyWebSigner } from "./signer.js";
 export type * from "./solanaSigner.js";
 export type * from "./types.js";
 export { AlchemySignerStatus } from "./types.js";
+export { SolanaSigner } from "./solanaSigner.js";