+++ site/sidebar/reference/account-kit/signer.ts
@@ -18,10 +18,6 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
         text: "AlchemySignerWebClient",
         items: [
           {
-            text: "addMfa",
-            link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/addMfa",
-          },
-          {
             text: "completeAuthWithBundle",
             link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/completeAuthWithBundle",
           },
@@ -42,10 +38,6 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
             link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/exportWallet",
           },
           {
-            text: "getMfaFactors",
-            link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/getMfaFactors",
-          },
-          {
             text: "initEmailAuth",
             link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/initEmailAuth",
           },
@@ -62,24 +54,12 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
             link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/oauthWithRedirect",
           },
           {
-            text: "removeMfa",
-            link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/removeMfa",
-          },
-          {
             text: "submitOtpCode",
             link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/submitOtpCode",
           },
           {
             text: "targetPublicKey",
             link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/targetPublicKey",
-          },
-          {
-            text: "validateMultiFactors",
-            link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/validateMultiFactors",
-          },
-          {
-            text: "verifyMfa",
-            link: "/reference/account-kit/signer/classes/AlchemySignerWebClient/verifyMfa",
           },
         ],
       },
@@ -114,10 +94,6 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
           {
             text: "disconnect",
             link: "/reference/account-kit/signer/classes/BaseAlchemySigner/disconnect",
-          },
-          {
-            text: "experimental_toSolanaSigner",
-            link: "/reference/account-kit/signer/classes/BaseAlchemySigner/experimental_toSolanaSigner",
           },
           {
             text: "exportWallet",
@@ -174,6 +150,10 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
           {
             text: "signTypedData",
             link: "/reference/account-kit/signer/classes/BaseAlchemySigner/signTypedData",
+          },
+          {
+            text: "toSolanaSigner",
+            link: "/reference/account-kit/signer/classes/BaseAlchemySigner/toSolanaSigner",
           },
           {
             text: "toViemAccount",
@@ -272,6 +252,31 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
           {
             text: "constructor",
             link: "/reference/account-kit/signer/classes/OauthCancelledError/constructor",
+          },
+        ],
+      },
+      {
+        text: "SolanaSigner",
+        items: [
+          {
+            text: "addSignature",
+            link: "/reference/account-kit/signer/classes/SolanaSigner/addSignature",
+          },
+          {
+            text: "addSponsorship",
+            link: "/reference/account-kit/signer/classes/SolanaSigner/addSponsorship",
+          },
+          {
+            text: "constructor",
+            link: "/reference/account-kit/signer/classes/SolanaSigner/constructor",
+          },
+          {
+            text: "createTransfer",
+            link: "/reference/account-kit/signer/classes/SolanaSigner/createTransfer",
+          },
+          {
+            text: "signMessage",
+            link: "/reference/account-kit/signer/classes/SolanaSigner/signMessage",
           },
         ],
       },